Fix argument count when using wasp shell routines
wasp_make_argv allocates an array of pointers using the
count passed to it. It then iterates over the list storing
pointers in the array. Since the list is one element greater
than the count due to the size being computed before the
executable name is prepended this results in an out of bounds
write.